just some few updates on the frontend,store and reducers

+++ graphql_files/graphql_post_root_values.js
@@ -1,52 +1,41 @@
-`type post{
-    author:user
-    id:ID!
-    post_title:String
-    post_content:String
-    post_tag:String
-  }
- 
-  
-  type Query{
-    query_post:[post]
-  
-  }
 
-  input create_post_input{
-    author_id:ID
-    post_title:String
-    post_content:String
-    post_tag:String
-  }
-  input edit_post_input{
-    id:ID
-    post_content:String
-  }
-  
-  type Mutation {
-    createUser(user_details:create_user_input): user
-    createPost(input:create_post_input):post
-    editPost(input:edit_post_input):post
-    deletePost(input:ID):post
-  }`
 
 const post = require("../models/post");
   let Post = require("../models/post");
   let {get_Author}= require("./helper_func")
   module.exports = {
-      query_post:async()=>{
+      query_post:async({input})=>{
+        console.log("yup")
+        console.log(input);
           try{
-            let posts = await Post.find({})
-            console.log(typeof posts[0]._doc.author)
-            return posts.map(post=>({...post._doc,
-                                    id:post._doc._id,
-                                    author:get_Author.bind(this,post._doc.author)
-                                }))
+                console.log(input.id);
+                let posts = await Post.findById(input.id)
+                console.log(typeof posts._doc.author)
+                return {...posts._doc,id:post._doc._id,author:get_Author.bind(this,post._doc.author)}
           }catch(err){
               console.log("there was an error in the graphql query post route")
           }
           
       },
+      query_posts:async(req)=>{
+          console.log(req)
+        try{
+          let posts = await Post.find({})
+          console.log(typeof posts[0]._doc.author)
+          let temp_data = posts.map(post=>({...post._doc,
+                                  id:post._doc._id,
+                                  author:get_Author.bind(this,post._doc.author)
+                              }))
+        //   temp_data.post_tag = temp_data.post_tag.join("");
+        let temp = temp_data.map(data=>({...data,post_tag:data.post_tag.join("")}))
+        console.log(temp)
+        
+          return temp;
+        }catch(err){
+            console.log("there was an error in the graphql query post route")
+        }
+        
+    },
       createPost: async ({input})=>{
           try{
               let new_post;
@@ -88,9 +77,12 @@ const post = require("../models/post");
         }
 
     },
-      deletePost: async ({input})=>{
+      deletePost: async ({input},{user})=>{
         try{
             let delete_post;
+            if(user.id != input){
+               throw "error"
+            }
             delete_post = await Post.findByIdAndDelete(input)
             if(!delete_post){
                 throw new Error("error");
@@ -101,6 +93,7 @@ const post = require("../models/post");
         }catch(err){
             console.log("there was an error in the delete post graphql route")
             console.log(err)
+            return null
         }
 
     }
